harness: absorb commons walk into mcp
